+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 - [Autofill]: added example to the Demo project
 
 - [Common]: added support to initialize `Language` with the `locale` object.
+- [Common]: added support to pass `Locale` object to the `SearchOptions`. Language code will be used for a search request if present.
 
 ## [1.0.0-beta.37] - 2022-10-14